Rename lakehouse monitor to quality monitor
Deprecate resource instead of breaking rename

fix

Apply suggestions from code review
