+++ internal/acceptance/quality_monitor_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 )
 
-var commonPartLakehouseMonitoring = `resource "databricks_catalog" "sandbox" {
+var commonPartQualityMonitoring = `resource "databricks_catalog" "sandbox" {
 	name         = "sandbox{var.STICKY_RANDOM}"
 	comment      = "this catalog is managed by terraform"
 	properties = {
@@ -46,16 +46,16 @@ resource "databricks_sql_table" "myInferenceTable" {
 
 `
 
-func TestUcAccLakehouseMonitor(t *testing.T) {
+func TestUcAccQualityMonitor(t *testing.T) {
 	if os.Getenv("GOOGLE_CREDENTIALS") != "" {
-		t.Skipf("databricks_lakehouse_monitor resource is not available on GCP")
+		t.Skipf("databricks_quality_monitor resource is not available on GCP")
 	}
 	unityWorkspaceLevel(t, step{
-		Template: commonPartLakehouseMonitoring + `
+		Template: commonPartQualityMonitoring + `
 
-			resource "databricks_lakehouse_monitor" "testMonitorInference" {
+			resource "databricks_quality_monitor" "testMonitorInference" {
 				table_name = databricks_sql_table.myInferenceTable.id
-				assets_dir = "/Shared/provider-test/databricks_lakehouse_monitoring/${databricks_sql_table.myInferenceTable.name}"
+				assets_dir = "/Shared/provider-test/databricks_quality_monitoring/${databricks_sql_table.myInferenceTable.name}"
 				output_schema_name = databricks_schema.things.id
 				inference_log  {
 				  granularities = ["1 day"]
@@ -79,9 +79,9 @@ func TestUcAccLakehouseMonitor(t *testing.T) {
 				}
 			}
 
-			resource "databricks_lakehouse_monitor" "testMonitorTimeseries" {
+			resource "databricks_quality_monitor" "testMonitorTimeseries" {
 				table_name = databricks_sql_table.myTimeseries.id
-				assets_dir = "/Shared/provider-test/databricks_lakehouse_monitoring/${databricks_sql_table.myTimeseries.name}"
+				assets_dir = "/Shared/provider-test/databricks_quality_monitoring/${databricks_sql_table.myTimeseries.name}"
 				output_schema_name = databricks_schema.things.id
 				time_series  {
 				  granularities = ["1 day"]
@@ -102,9 +102,9 @@ func TestUcAccLakehouseMonitor(t *testing.T) {
 				}
 			}
 
-			resource "databricks_lakehouse_monitor" "testMonitorSnapshot" {
+			resource "databricks_quality_monitor" "testMonitorSnapshot" {
 				table_name = databricks_sql_table.mySnapshot.id
-				assets_dir = "/Shared/provider-test/databricks_lakehouse_monitoring/${databricks_sql_table.myTimeseries.name}"
+				assets_dir = "/Shared/provider-test/databricks_quality_monitoring/${databricks_sql_table.myTimeseries.name}"
 				output_schema_name = databricks_schema.things.id
 				snapshot  {
 				} 
@@ -113,15 +113,15 @@ func TestUcAccLakehouseMonitor(t *testing.T) {
 	})
 }
 
-func TestUcAccUpdateLakehouseMonitor(t *testing.T) {
+func TestUcAccUpdateQualityMonitor(t *testing.T) {
 	if os.Getenv("GOOGLE_CREDENTIALS") != "" {
-		t.Skipf("databricks_lakehouse_monitor resource is not available on GCP")
+		t.Skipf("databricks_quality_monitor resource is not available on GCP")
 	}
 	unityWorkspaceLevel(t, step{
-		Template: commonPartLakehouseMonitoring + `
-			resource "databricks_lakehouse_monitor" "testMonitorInference" {
+		Template: commonPartQualityMonitoring + `
+			resource "databricks_quality_monitor" "testMonitorInference" {
 				table_name = databricks_sql_table.myInferenceTable.id
-				assets_dir = "/Shared/provider-test/databricks_lakehouse_monitoring/${databricks_sql_table.myInferenceTable.name}"
+				assets_dir = "/Shared/provider-test/databricks_quality_monitoring/${databricks_sql_table.myInferenceTable.name}"
 				output_schema_name = databricks_schema.things.id
 				inference_log  {
 				  granularities = ["1 day"]
@@ -133,10 +133,10 @@ func TestUcAccUpdateLakehouseMonitor(t *testing.T) {
 			}
 		`,
 	}, step{
-		Template: commonPartLakehouseMonitoring + `
-		resource "databricks_lakehouse_monitor" "testMonitorInference" {
+		Template: commonPartQualityMonitoring + `
+		resource "databricks_quality_monitor" "testMonitorInference" {
 			table_name = databricks_sql_table.myInferenceTable.id
-			assets_dir = "/Shared/provider-test/databricks_lakehouse_monitoring/${databricks_sql_table.myInferenceTable.name}"
+			assets_dir = "/Shared/provider-test/databricks_quality_monitoring/${databricks_sql_table.myInferenceTable.name}"
 			output_schema_name = databricks_schema.things.id
 			inference_log  {
 			  granularities = ["1 hour"]
